+++ API/GameEngineContents/Vector2D.cpp
@@ -11,7 +11,7 @@ void Vector2D::DebugVectorRender(GameEngineActor* _Actor)
 	TCHAR Buffer[30] = "";
 	sprintf_s(Buffer, "Pos { %d, %d } ", _Actor->GetPosition().ix(), _Actor->GetPosition().iy());
 
-	TextOutA(hdc, _Actor->GetCameraEffectPosition().ix(), _Actor->GetCameraEffectPosition().iy() - 50, Buffer, strlen(Buffer));
+	TextOutA(hdc, _Actor->GetCameraEffectPosition().ix(), _Actor->GetCameraEffectPosition().iy() - 50, Buffer, static_cast<int>(strlen(Buffer)));
 }
 
 Vector2D::Vector2D()
